Check for, and squelch, signed vs. unsigned comparison errors.
Use -Wsign-compare if the compiler supports it.

If we're comparing an unsigned value with a signed value, and we've
already determined that the signed value is >= 0, just cast it to an
unsigned type.

Use 0xffffffffU instead of -1 as the "unknown"/"unset"/etc. value for
unsigned variables and fields.

Assign the result of str2tok() to an int, not a u_int, as it can return
-1.

Declare some variables and fields unsigned if they don't need to be
signed.  Make sure some arguments passed into "set" functions are
non-negative (and otherwise not invalid).

In the BPF optimizer, cast the "constant" fields of a struct block to an
unsigned type - the actual constant field of a BPF instruction is
unsigned.

Also, cast away one warning from MSVC.

+++ bpf/net/bpf_filter.c
@@ -734,7 +734,7 @@ bpf_validate(f, len)
 #if defined(KERNEL) || defined(_KERNEL)
 				if (from + p->k < from || from + p->k >= len)
 #else
-				if (from + p->k >= len)
+				if (from + p->k >= (u_int)len)
 #endif
 					return 0;
 				break;
@@ -742,7 +742,7 @@ bpf_validate(f, len)
 			case BPF_JGT:
 			case BPF_JGE:
 			case BPF_JSET:
-				if (from + p->jt >= len || from + p->jf >= len)
+				if (from + p->jt >= (u_int)len || from + p->jf >= (u_int)len)
 					return 0;
 				break;
 			default:
